% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcion.R
\name{retrieve_answer}
\alias{retrieve_answer}
\title{Show one of the 6 exercise of Assignment 1}
\usage{
retrieve_answer(n)
}
\arguments{
\item{n}{a numeric value from 1 to 6}
}
\value{
The script of the exercise
}
\description{
Show one of the 6 exercise of Assignment 1
}
