% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcion.R
\name{retrieve_answer}
\alias{retrieve_answer}
\title{retrieve_answer:Show the solution of one of the 6 exercise of Assignment 1}
\usage{
retrieve_answer(n)
}
\arguments{
\item{n}{a numeric value from 1 to 6}
}
\value{
The solution of the chosen exercise
}
\description{
The "retrieve_answer" is a function that shows the solution of one of the 6 assigned exercises corresponding to chapter 5 "Data Transformation" of R for Data Science where:
1 = 5.2.4 Exercises: items 1, and 2
2 = 5.3.1 Exercises: all items
3 = 5.4.1 Exercises: items 2, 3, and 4
4 = 5.5.2 Exercises: items 1, and 2
5 = 5.6.7 Exercises: item 1
6 = 5.7.1 Exercises: item 2#'
}
